Django_RiotAPI/requirements.txt
asgiref==3.2.7
boto3==1.14.11
botocore==1.17.11
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.7
django-allauth==0.42.0
django-storages==1.9.1
docutils==0.15.2
idna==2.9
jmespath==0.10.0
oauthlib==3.1.0
Pillow==7.1.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9